Refactor: move fillInput and getErrorText functions to steps/genericSteps.ts
